small fixes

+++ database/migrations/2023_06_10_141944_create_accounts_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained();
             $table->string('number')->unique();
             $table->string('currency');
-            $table->float('balance')->default(100.00);
-            $table->float('invested_amount')->default(0.00);
+            $table->decimal('balance', 10,2)->default(100.00);
+            $table->decimal('invested_amount', 10,2)->default(0.00);
             $table->string('name');
             $table->timestamps();
         });
