+++ database/migrations/2023_06_10_142006_create_transactions_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('currency_to');
             $table->float('amount');
             $table->float('amount_converted');
-            $table->float('exchange_rate');
+            $table->decimal('exchange_rate', 6);
             $table->timestamps();
         });
     }
